// src/types/hass.type.ts
export enum Hass_Device_Class {
  APPARENT_POWER = "apparent_power", // Apparent power in VA.
  AQI = "aqi", // Air Quality Index
  BATTERY = "battery", // Percentage of battery that is left
  CARBON_DIOXIDE = "carbon_dioxide", // Carbon Dioxide in CO2 (Smoke)
  CARBON_MONOXIDE = "carbon_monoxide", // Carbon Monoxide in CO (Gas CNG/LPG)
  CURRENT = "current", // Current in A
  DATE = "date", // Date string (ISO 8601)
  DISTANCE = "distance", // Generic distance in km, m, cm, mm, mi, yd, or in
  DURATION = "duration", // Duration in days, hours, minutes or seconds
  ENERGY = "energy", // Energy in Wh, kWh or MWh
  FREQUENCY = "frequency", // Frequency in Hz, kHz, MHz or GHz
  GAS = "gas", // Gasvolume in m³ or ft³
  HUMIDITY = "humidity", // Percentage of humidity in the air
  ILLUMINANCE = "illuminance", // The current light level in lx or lm
  MOISTURE = "moisture", // Percentage of water in a substance
  MONETARY = "monetary", // The monetary value
  NITROGEN_DIOXIDE = "nitrogen_dioxide", // Concentration of Nitrogen Dioxide in µg/m³
  NITROGEN_MONOXIDE = "nitrogen_monoxide", // Concentration of Nitrogen Monoxide in µg/m³
  NITROUS_OXIDE = "nitrous_oxide", // Concentration of Nitrous Oxide in µg/m³
  OZONE = "ozone", // Concentration of Ozone in µg/m³
  PM1 = "pm1", // Concentration of particulate matter less than 1 micrometer in µg/m³
  PM10 = "pm10", // Concentration of particulate matter less than 10 micrometers in µg/m³
  PM25 = "pm25", // Concentration of particulate matter less than 2.5 micrometers in µg/m³
  POWER_FACTOR = "power_factor", // Power factor in %
  POWER = "power", // Power in W or kW
  PRESSURE = "pressure", // Pressure in Pa, kPa, hPa, bar, cbar, mbar, mmHg, inHg, or psi
  REACTIVE_POWER = "reactive_power", // Reactive power in var
  SIGNAL_STRENGTH = "signal_strength", // Signal strength in dB or dBm
  SPEED = "speed", // Generic speed in ft/s, in/d, in/h, km/h, kn, m/s, mph, or mm/d
  SULPHUR_DIOXIDE = "sulphur_dioxide", // Concentration of sulphur dioxide in µg/m³
  TEMPERATURE = "temperature", // Temperature in °C or °F
  TIMESTAMP = "timestamp", // Datetime object or timestamp string (ISO 8601)
  VOLATILE_ORGANIC_COMPOUNDS = "volatile_organic_compounds", // Concentration of volatile organic compounds in µg/m³
  VOLTAGE = "voltage", // Voltage in V
  VOLUME = "volume", // Generic volume in L, mL, gal, fl. oz., m³, or ft³
  WEIGHT = "weight", // Generic mass in kg, g, mg, µg, oz, or lb
}